requirements.txt
torch==2.5.1
torchvision==0.20.1
torchaudio==2.5.1
pillow==10.4.0
numpy==2.1.1
transformers==4.48.1
jupyterlab==4.2.5
datasets==3.0.0
matplotlib==3.9.2
scikit-learn==1.5.2
scipy==1.14.1
seaborn==0.13.2
sentencepiece==0.2.0
protobuf==5.28.2
rich==13.9.2
soundfile==0.12.1
einops==0.8.0
wandb==0.18.7
librosa==0.10.2.post1
av==13.1.0
jsonlines==4.0.0
evaluate==0.4.3
accelerate==1.3.0
rouge_score
timm==1.0.15
sentence-transformers==3.4.1
scikit-image==0.24.0
